ml-pipeline/requirements.txt
# Core ML Libraries
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# Deep Learning
torch==2.1.2
transformers==4.36.2

# Database
neo4j==5.15.0
redis==5.0.1

# API Framework
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.2
hypothesis==6.92.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
